estagiarios-platform/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
websockets==12.0
pydantic==2.11.7
python-dotenv==1.0.0
twilio==8.10.0
